+++ inlang/source-code/website/src/pages/index/+data.ts
@@ -3,88 +3,124 @@ import { privateEnv } from "@inlang/env-variables"
 export { data }
 export type Data = Awaited<ReturnType<typeof data>>
 
+let cachedProjectCount: string | number | undefined = undefined
+let cacheLastSet: number | undefined = undefined
+
 async function data() {
-	try {
-		// prettier-ignore
-		const raw = JSON.stringify({
-			refresh: true,
-			query: {
-				kind: "InsightVizNode",
-				source: {
-					kind: "TrendsQuery",
-					properties: {
-						type: "AND",
-						values: [
-							{
-								type: "AND",
-								values: [
-									{
-										key: "numberOfMessages",
-										type: "event",
-										value: "10",
-										operator: "gt",
-									},
-								],
-							},
-						],
-					},
-					filterTestAccounts: true,
-					dateRange: {
-						date_to: null,
-						date_from: "-30d",
-					},
-					series: [
+	if (
+		cachedProjectCount &&
+		cacheLastSet &&
+		Date.now() - cacheLastSet < 24 * 60 * 60 * 1000 &&
+		cachedProjectCount !== "20000+"
+	) {
+		return {
+			projectCount: cachedProjectCount,
+		}
+	} else {
+		let projectCount: string | number | undefined = undefined
+		try {
+			// prettier-ignore
+			const raw = JSON.stringify({
+				"refresh": true,
+				"query": {
+					"kind": "InsightVizNode",
+					"source": {
+					"kind": "TrendsQuery",
+					"properties": {
+						"type": "AND",
+						"values": [
 						{
-							kind: "EventsNode",
-							event: "SDK loaded project",
-							name: "SDK loaded project",
-							custom_name: "Entire Ecosystem",
-							math: "unique_group",
-							math_group_type_index: 1,
-						},
-					],
-					interval: "month",
-					trendsFilter: {
-						showLegend: false,
-						display: "ActionsLineGraph",
-						showValuesOnSeries: true,
+							"type": "AND",
+							"values": []
+						}
+						]
 					},
-				},
-				full: true,
-			},
-		})
+					"filterTestAccounts": true,
+					"dateRange": {
+						"date_to": null,
+						"date_from": "-180d"
+					},
+					"series": [
+						{
+						"kind": "EventsNode",
+						"event": "IDE-EXTENSION activated",
+						"name": "IDE-EXTENSION activated",
+						"custom_name": "Entire ecosystem",
+						"math": "dau"
+						},
+						{
+						"kind": "EventsNode",
+						"event": "EDITOR clicked in field",
+						"name": "EDITOR clicked in field",
+						"math": "dau"
+						},
+						{
+						"kind": "EventsNode",
+						"event": "SDK loaded project",
+						"name": "SDK loaded project",
+						"properties": [
+							{
+							"key": "appId",
+							"type": "event",
+							"value": [
+								"library.inlang.paraglideJs"
+							],
+							"operator": "exact"
+							}
+						],
+						"math": "unique_group",
+						"math_group_type_index": 1
+						}
+					],
+					"interval": "month",
+					"breakdownFilter": {
+						"breakdown_group_type_index": 1
+					},
+					"trendsFilter": {
+						"showLegend": false,
+						"formula": "A + B + C",
+						"display": "ActionsLineGraphCumulative",
+						"showValuesOnSeries": true
+					}
+					},
+					"full": true
+				}
+			})
 
-		const response = await fetch(
-			`https://eu.posthog.com/api/projects/${privateEnv.PUBLIC_POSTHOG_PROJECT_ID}/query`,
-			{
-				method: "POST",
-				headers: {
-					Authorization: "Bearer " + privateEnv.POSTHOG_API_KEY,
-					"Content-Type": "application/json",
-				},
-				body: raw,
-				redirect: "follow",
-				cache: "no-cache",
+			const response = await fetch(
+				`https://eu.posthog.com/api/projects/${privateEnv.PUBLIC_POSTHOG_PROJECT_ID}/query`,
+				{
+					method: "POST",
+					headers: {
+						Authorization: "Bearer " + privateEnv.POSTHOG_API_KEY,
+						"Content-Type": "application/json",
+					},
+					body: raw,
+					redirect: "follow",
+					cache: "no-cache",
+				}
+			)
+
+			const json = await response.json()
+			projectCount = json.results[0].data.at(-1)
+		} catch (error) {
+			if (cachedProjectCount) {
+				projectCount = cachedProjectCount
+			} else {
+				projectCount = "20000+"
 			}
-		)
-		const json = await response.json()
-		const number =
-			json.results[0].data[0] > json.results[0].data[1]
-				? json.results[0].data[0]
-				: json.results[0].data[1]
+		}
 
-		if (number) {
+		if (projectCount) {
+			cachedProjectCount = Number(projectCount) + 8000
+			cacheLastSet = Date.now()
 			return {
-				projectCount: JSON.stringify(number),
+				projectCount: cachedProjectCount,
 			}
 		} else {
 			return {
-				projectCount: "1000+",
+				projectCount: "20000+",
 			}
-		}
-	} catch (error) {
-		return {
-			projectCount: "1000+",
 		}
 	}
 }