+++ inlang/source-code/plugins/json/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @inlang/plugin-json
 
+## 5.1.49
+
+### Patch Changes
+
+- Updated dependencies [2a5645c]
+  - @inlang/sdk@0.35.7
+
+## 5.1.48
+
+### Patch Changes
+
+- Updated dependencies [9d2aa1a]
+  - @inlang/sdk@0.35.6
+
 ## 5.1.47
 
 ### Patch Changes