+++ inlang/source-code/server/CHANGELOG.md
@@ -4,6 +4,26 @@
 
 ### Patch Changes
 
+- @inlang/badge@0.7.34
+- @inlang/plugin-message-format@2.2.0
+- @inlang/rpc@0.3.29
+- @inlang/telemetry@0.3.44
+- @inlang/website@null
+
+## null
+
+### Patch Changes
+
+- @inlang/plugin-message-format@2.2.0
+- @inlang/website@null
+- @inlang/badge@0.7.33
+- @inlang/rpc@0.3.28
+- @inlang/telemetry@0.3.43
+
+## null
+
+### Patch Changes
+
 - @inlang/badge@0.7.32
 - @inlang/plugin-message-format@2.2.0
 - @inlang/rpc@0.3.27